+++ CHANGELOG.md
@@ -11,9 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `opentelemetry-instrumentation-kafka-python` Instrument temporary fork, kafka-python-ng
   inside kafka-python's instrumentation
-  ([#2537](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2537)))
+  ([#2537](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2537))
 
 ## Breaking changes
+
+- `opentelemetry-bootstrap` Remove `opentelemetry-instrumentation-aws-lambda` from the defaults instrumentations
+  ([#2537](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2537))
 
 ## Fixed